pyproject.toml
[tool.poetry]
name = "noverde_backend"
version = "0.1.0"
description = "Desafio backend"
authors = ["Fred Campos <fred.tecnologia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
transitions = "^0.8.1"
django = "^3.1"
django-celery-results = "^1.2.1"
djangorestframework = "^3.11.1"
redis = "^3.5.3"
psycopg2 = "^2.8.5"
schema = "^0.7.3"
validate-docbr = "^1.7.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
bump2version = "^0.5.11"


cohesion = "^1.0.0"
darglint = "^1.3.0"
dlint = "^0.10.3"
epc = "^0.0.5"
flake8-cognitive-complexity = "^0.0.2"
flake8-executable = "^2.0.3"
flake8-expression-complexity = "^0.0.7"
flake8-functions = "^0.0.4"
flake8-logging-format = "^0.6.0"
flake8-mock = "^0.3"
flake8-pie = "^0.5.0"
flake8-pytest-style = "^1.1.1"
flake8-variables-names = "^0.0.3"
importmagic3 = "^0.2.0"
ipython = "^7.14.0"
monkeytype = "^20.4.2"
mypy = "^0.770"
pip = "^20.1.0"
py-make = "^0.1.1"
pylint = "^2.5.2"

pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-mock = "^3.1.0"

sphinx = "^1.8.5"
tox = "^3.14.6"
watchdog = "^0.9.0"
wemake-python-styleguide = "^0.14.0"
mixer = "^6.1.3"
pytest-django = "^3.9.0"

[tool.black]
line-length = 100
skip-string-normalization = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
